Calculate most similar book sections
We want to provide an excerpt of relevant book sections to OpenAI to
provide context for the completion. To do this, given we already have
embeddings computed for the book by page, we can compute similarity with
the question by using a dot/inner product (because OpenAI embedding vectors
are normalized, this is equivalent to cosine, i.e. the closer we are to
1, the more similar 2 vectors are
https://platform.openai.com/docs/guides/embeddings/limitations-risks)

BookSimilarity#book_sections_sorted_by_relevance just returns the array
of book section titles and embeddings sorted by similarity to question