Limit the amount of relevant book context
We need to limit the number of tokens we send to OpenAI's completion API
According to https://platform.openai.com/docs/api-reference/completions/create#completions/create-max_tokens
```
The token count of your prompt plus max_tokens cannot exceed the
model's context length. Most models have a context length of 2048 tokens
(except for the newest models, which support 4096).
```

We are using `text-davinci-003` for completions, which according to
https://platform.openai.com/docs/models/gpt-3, has a limit of 4000
tokens.

The hardcoded prompt in `prompt_templater.rb` has 881 tokens
(`Tokenizers.from_pretrained("gpt2").encode(PromptTemplater.new.construct_prompt("")).tokens.count`),
and we are limiting to 500 tokens here, so there should be plenty of
room for the user's question + OpenAI's completion.

One caveat is that truncating the book text to fit the 500 token limit
has some imperfections due to inability to easily truncate a string up
to an n-th token. Instead we choose to slice by characters, which should
keep us below or at the limit.