Remove non ASCII characters from book contents
pdf-reader is including https://www.compart.com/en/unicode/U+25AF when
parsing text, which appears to be messing up OpenAI embeddings, so strip
them out using String#encode (followed tips in
https://stackoverflow.com/a/9420531 but updated for Ruby 3)

I found this out because the question I've been testing with the
lecture1.pdf is "How much data does the Large Hadron Collider produce?".
Without the book context, OpenAI responds with 25 petabytes, but using
the book context it'll answer with 15 petabytes. For some reason the
page with this sentence was not being ranked the most similar to the
question, so I suspected it was something to do with these strange
U+25AF characters.